Fix issue with MedianCDFQuantileScorer
Before, the scorer was not able to handle numpy object types directly. However, GCM often uses the object dtype to ensure support of mixing categorical and float values. This fixes the handling of object dtypes by explicitly converting them to floats first.

+++ dowhy/gcm/anomaly_scorers.py
@@ -43,13 +43,13 @@ class MedianCDFQuantileScorer(AnomalyScorer):
         if (X.ndim == 2 and X.shape[1] > 1) or X.ndim > 2:
             raise ValueError("The MedianCDFQuantileScorer currently only supports one-dimensional data!")
 
-        self._distribution_samples = X.reshape(-1)
+        self._distribution_samples = X.reshape(-1).astype(float)
 
     def score(self, X: np.ndarray) -> np.ndarray:
         if self._distribution_samples is None:
             raise ValueError("Scorer has not been fitted!")
 
-        X = shape_into_2d(X)
+        X = shape_into_2d(X.astype(float))
 
         equal_samples = np.sum(np.isclose(X, self._distribution_samples, rtol=0, atol=0, equal_nan=True), axis=1)
         greater_samples = np.sum(X > self._distribution_samples, axis=1) + equal_samples / 2
